mm-hie-backend/requirements.txt
fastapi
uvicorn
pdfplumber
torch
transformers
torchvision
boto3
scikit-learn
opencv-python
numpy
pandas
python-multipart
Pillow
shap
lime
reportlab
SQLAlchemy
psycopg2-binary
alembic
python-dotenv
grad-cam
rank_bm25

# RAG/LLM dependencies
sentence-transformers
faiss-cpu
llama-cpp-python
pytest
httpx
datasets
peft
bitsandbytes
accelerate
scipy
